Diploma/requirements.txt
plotly == 5.10.0
pygame == 2.1.2
geopy == 2.2.0
geopandas == 0.11.1
dash == 2.6.1